gtk: generate ApplicationBuilder manually
fixes #143
